feat: 使用 fast-steal 库

+++ src/get_url_info.rs
@@ -10,7 +10,7 @@ use sanitize_filename;
 #[derive(Clone)]
 #[allow(dead_code)]
 pub struct UrlInfo {
-    pub file_size: usize,
+    pub file_size: u64,
     pub file_name: String,
     pub supports_range: bool,
     pub final_url: String,
@@ -18,7 +18,7 @@ pub struct UrlInfo {
     pub last_modified: Option<String>,
 }
 
-fn get_file_size(headers: &HeaderMap, status: &StatusCode) -> usize {
+fn get_file_size(headers: &HeaderMap, status: &StatusCode) -> u64 {
     if *status == StatusCode::PARTIAL_CONTENT {
         headers
             .get(header::CONTENT_RANGE)
